+++ src/main/java/com/git/Nog022/MeetHub/dto/RoomDTO.java
@@ -5,7 +5,7 @@ import java.util.List;
 public record RoomDTO(
         String name,
         Integer capacity,
-        Integer localId,
+        Long localId,
         List<String> resources
 ) {
 }
